Interactive Web
This project was to create a web sight that accepted a date as an input and then pulled data for that input. 
We were provided the Data set. 
The Data was on UFO Sightings. 
Was a fun project to create a web sight that accepted an input and searched through Data and then output a table.